refactor firebaseApp into multiple files
